+++ familyhelper-webapi-node/src/main/java/com/dwarfeng/familyhelper/webapi/node/configuration/settingrepo/BeanTransformerConfiguration.java
@@ -3,9 +3,11 @@ package com.dwarfeng.familyhelper.webapi.node.configuration.settingrepo;
 import com.dwarfeng.settingrepo.sdk.bean.entity.FastJsonFormatterSupport;
 import com.dwarfeng.settingrepo.sdk.bean.entity.FastJsonSettingCategory;
 import com.dwarfeng.settingrepo.sdk.bean.entity.FastJsonSettingNode;
+import com.dwarfeng.settingrepo.sdk.bean.entity.FastJsonTextNode;
 import com.dwarfeng.settingrepo.stack.bean.entity.FormatterSupport;
 import com.dwarfeng.settingrepo.stack.bean.entity.SettingCategory;
 import com.dwarfeng.settingrepo.stack.bean.entity.SettingNode;
+import com.dwarfeng.settingrepo.stack.bean.entity.TextNode;
 import com.dwarfeng.subgrade.impl.bean.DozerBeanTransformer;
 import com.dwarfeng.subgrade.stack.bean.BeanTransformer;
 import org.dozer.Mapper;
@@ -34,5 +36,10 @@ public class BeanTransformerConfiguration {
     @Bean("settingrepo.formatterSupportBeanTransformer")
     public BeanTransformer<FormatterSupport, FastJsonFormatterSupport> formatterSupportBeanTransformer() {
         return new DozerBeanTransformer<>(FormatterSupport.class, FastJsonFormatterSupport.class, mapper);
+    }
+
+    @Bean("settingrepo.textNodeBeanTransformer")
+    public BeanTransformer<TextNode, FastJsonTextNode> textNodeBeanTransformer() {
+        return new DozerBeanTransformer<>(TextNode.class, FastJsonTextNode.class, mapper);
     }
 }